Just for fun: Invert two factors and see what happens.
This must have been an experiment when the maths following logic failed for some (probably different) reason. This is supported by the fact that I didn't note what I wanted to achieve with these changes or why this seemed sensible. Probably this just was trial and error but didn't find the issue. I remember that something bugged me, because the logical conclusions didn't work out. Anyway, this is legacy but nevertheless better is tracked in a commit, which is the reason why this commit exists.