The quest of the magic number 1.3648.
Calculating it failed, but not using it lead to worse results.
Blenderartist reC's friend might know the answer! Either we ask for
help or we try to figure where this magic number can be derived from
ourselves (if I remember correctly then the number may have been
experimentally determined for the A4 ratio).

In the code my trigonometric solution (that didn't quite print to scale! if someone wants to help us blender engineers and artists out we'd be grateful) can be found.

This delivers a working version for landscape mode. Portrait is still buggy.